refactor: restructure runtime code
