+++ .eslintignore
@@ -1,3 +1,2 @@
 /node_modules
-/lib/url/escape.js
-/lib/css-base.js
+/lib/runtime
