+++ lib/loader.js
@@ -113,7 +113,7 @@ module.exports = function loader(content, map) {
       if (query.url !== false && result.urlItems.length > 0) {
         urlEscapeHelper = `var escape = require(${loaderUtils.stringifyRequest(
           this,
-          require.resolve('./url/escape.js')
+          require.resolve('./runtime/escape.js')
         )});\n`;
 
         cssAsString = cssAsString.replace(result.urlItemRegExpG, (item) => {
@@ -184,7 +184,7 @@ module.exports = function loader(content, map) {
         null,
         `${urlEscapeHelper}exports = module.exports = require(${loaderUtils.stringifyRequest(
           this,
-          require.resolve('./css-base.js')
+          require.resolve('./runtime/api.js')
         )})(${sourceMap});\n` +
           `// imports\n${importJs}\n\n` +
           `// module\n${moduleJs}\n\n` +
