+++ lib/runtime/escape.js
